client_requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
httpx